Improved Readiness Checks (#45)
Adds better methods for determining system readiness

The use of `waitForReady` was proving very brittle. often it would either exit too early or just exit immediately if the client ran into a problem. This replaces the usage of waitForReady with some shell functions.

By default, the new `isReady` function will use a similar method to the old `waitForReady` command. This new method gets the output of `lna` and filters out the table header, next it filters out any bundles included in the `READINESS_EXCLUSIONS` variable, finally it counts the number of bundles left in the list. If the number of bundles is not `0` it returns an exit code of `1`. If the number of bundles in the list is equal to `0` it returns an exit code of `0`

This `isReady` function is used by the new `waitForReady` function which will just block further execution until `isReady` returns `0`

Further, the behavior of the `isReady` function can be altered by setting `EXPERIMENTAL_READINESS_CHECKS_ENABLED=true`. When this variable is present the `isReady` function will utilize the `fabric8-karaf-checks` provided endpoint to determine the readiness of the system instead of the `lna` based approach